Mediapipe cleanup (#1787)
* Split mediapipe configuration management from calculators
* Divide mediapipe graph definition and execution
* Improve return codes handling and logging
* Remove unnecessary calculators and files
* Ignore not found headers in cppclean
* Fix issues with not defining functions as static
JIRA:CVS-107512, CVS-107513
