+++ src/http_server.cpp
@@ -39,7 +39,7 @@ namespace ovms {
 
 namespace net_http = tensorflow::serving::net_http;
 
-const net_http::HTTPStatusCode http(const ovms::Status& status) {
+static const net_http::HTTPStatusCode http(const ovms::Status& status) {
     const std::unordered_map<const StatusCode, net_http::HTTPStatusCode> httpStatusMap = {
         {StatusCode::OK, net_http::HTTPStatusCode::OK},
         {StatusCode::OK_RELOADED, net_http::HTTPStatusCode::CREATED},
