+++ src/test/mediapipe/graphdummyadapterfull.pbtxt
@@ -30,7 +30,6 @@ node {
         key: "A"
         value: "a"
       }
-      config_path: "/ovms/src/test/mediapipe/config_standard_dummy.json"
     }
   }
 }
